requirements.txt
qdrant-client==1.7.0
openai==1.6.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.2
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
requests==2.31.0
